% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bHP.R
\docType{package}
\name{bHP}
\alias{bHP}
\alias{bHP-package}
\title{BoostedHP: A package for the boosted HP filter}
\description{
Iterated HP filtering by Phillips and Shi (2020)
}
\section{Boosted HP main function}{


BoostedHP is the main function
}

\references{
Phillips, Peter CB, and Zhentao Shi.
"Boosting: Why you can use the hp filter."
arXiv: 1905.00175, Cowles Foundation Discussion Paper No.2192, (2019).
}
