% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedHP.R
\name{BoostedHP}
\alias{BoostedHP}
\title{Title}
\usage{
BoostedHP(x, lambda = 1600, iter = TRUE, test_type = "none",
  sig_p = 0.05, Max_Iter = 100)
}
\arguments{
\item{Max_Iter}{}
}
\description{
Title
}
