% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRE.R
\docType{data}
\name{IRE}
\alias{IRE}
\title{\code{IRE} Ireland Annual GDP}
\format{An \code{\link{xts}} object of Real Exports of Goods and Services.
 \itemize{
 \item\strong{Release:} {Gross Domestic Product}
 \item\strong{Seasonal Adjustment:} {Seasonally Adjusted Annual Rate}
 \item\strong{Frequency:} {Quarterly}
 \item\strong{Units:} {Billions of Chained 2009 Dollars}
 \item\strong{Date Range:} {1947-01-01 to 2017-07-01}
 \item\strong{Last Updated} {2017-12-21 7:51 AM CST}
}}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/EXPGSC1.txt}
}
\usage{
data(EXPGSC1)
}
\description{
\code{IRE} Ireland Annual GDP
}
\section{Notes}{

BEA Account Code: A020RX

A Guide to the National Income and Product Accounts of the United States (NIPA)
- (http://www.bea.gov/national/pdf/nipaguid.pdf)
}

\examples{
data(IRE)
tail(IRE)
plot(IRE,grid.col = "white", col="blue")
}
\keyword{datasets}
