% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.bHP}
\alias{predict.bHP}
\title{Predict.bHP}
\usage{
\method{predict}{bHP}(x)
}
\arguments{
\item{x}{an object of class \code{bHP}}
}
\value{
the estimated trend component
}
\description{
Extract the final trend component for class \code{bHP}.
}
\examples{
lam <- 100 # tuning parameter for the annual data

data(IRE) # load the data 'IRE'

# raw HP filter
bx_HP <- BoostedHP(IRE, lambda = lam, iter = FALSE)

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter = TRUE, stopping = "adf", sig_p = 0.050)

# return the final trend component

predict(bx_HP)

predict(bx_ADF)
}
