% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bHP.R
\name{summary.bHP}
\alias{summary.bHP}
\title{Summary method for class 'bHP'}
\usage{
\method{summary}{bHP}(x, digit = 8)
}
\arguments{
\item{x}{an object of class "bHP"}

\item{digit}{controls the number of significant digits to print when printing numeric values. It is a suggestion only. Valid values are 1...22 with default 8. See the note in 'print.default' about values greater than 15.}
}
\value{
summary of raw dta, trend component, iterated number, p-value ...
}
\description{
table of the summary special for class 'bHP'.
}
\examples{
lam <- 100 # tuning parameter for the annaul data

data(IRE) # laod the data 'IRE'

# raw HP filter
bx_HP <- BoostedHP(x, lambda = lam, iter= FALSE)

# by BIC
bx_BIC <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "BIC")

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "adf")

# by none test type
# Iterated HP filter until Max_Iter and keep the path of BIC.

bx_none <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "none")

#--------- start to summary the content of bHP -----------------

summary(bx_ADF)
summary(bx_BIC)
summary(bx_none)
summary(bx_HP)

}
