% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bHP.R
\docType{package}
\name{bHP}
\alias{bHP}
\alias{bHP-package}
\title{BoostedHP: Package for the boosted HP filter}
\description{
Iterated HP filtering by Phillips and Shi (2019)
}
\details{
\code{BoostedHP} is the main function. It genereates a \code{bHP} object that can work with
generic methods.
}
\references{
Phillips, Peter CB, and Zhentao Shi.
"Boosting: Why you can use the hp filter."
arXiv: 1905.00175, Cowles Foundation Discussion Paper No.2192, (2019).
}
