% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bHP.R
\docType{package}
\name{bHP}
\alias{bHP}
\title{bHP: Package for the boosted HP filter}
\description{
The boosted HP filter by Phillips and Shi (2019)
}
\details{
\code{BoostedHP} is the main function. It generates a \code{bHP} object which can work with
generic methods.
}
\references{
Peter Phillips and Zhentao Shi, 2021: "Boosting: Why You Can Use the HP Filter," International Economic Review, 62(2), 521-570
}
