% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedHP.R
\name{BoostedHP}
\alias{BoostedHP}
\title{Boosting the Hodrick-Prescott Filter}
\usage{
BoostedHP(x, lambda = 1600, iter = TRUE, test_type = "BIC",
  sig_p = 0.05, Max_Iter = 100)
}
\arguments{
\item{x}{an object of class "bHP"}

\item{lambda}{the turning parameter, default value is 1600.}

\item{iter}{logical parameter, TRUE (default) is to conduct iterated HP-filter, FALSE is not.}

\item{test_type}{the type for creterion: none-iter, adf, BIC (default), none.}

\item{sig_p}{a threshold of the p-value below which the iteration will stop. default value is 0.050.}

\item{Max_Iter}{maximal number of iterations. The default value is 100.}
}
\value{
The function returns a list containing the following items
\item{cycle}{The cyclical component in the final iteration.}
\item{trend}{The trend component in the final iteration.}
\item{trend_hist}{The estimated trend in each iteration.}
\item{iter_num}{The total number of iterations when it stops.}
\item{IC_hist}{The path of the BIC through the iterations.}
\item{adf_p_hist}{The path of the ADF test p-value through the iterations}
}
\description{
all in one function of conducting iterated HP-filter for types: none-iter, adf, BIC, none.
}
\examples{
library(tseries)

data(IRE) # load the data 'IRE'
lam <- 100 # tuning parameter for the annaul data

# raw HP filter
bx_HP <- BoostedHP(IRE, lambda = lam, iter= FALSE)

# by BIC
bx_BIC <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "BIC")

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "adf")

# by none test type
# Iterated HP filter until Max_Iter and keep the path of BIC.

bx_none <- BoostedHP(IRE, lambda = lam, iter= TRUE, test_type = "none")

}
