% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedHP.R
\name{BoostedHP}
\alias{BoostedHP}
\title{Boosting the Hodrick-Prescott Filter}
\usage{
BoostedHP(
  x,
  lambda = 1600,
  iter = TRUE,
  stopping = "BIC",
  sig_p = 0.05,
  Max_Iter = 100
)
}
\arguments{
\item{x}{is a raw time series to be filtered.}

\item{lambda}{the turning parameter, default value is 1600,
as recommended by Hodrick and Prescott (1997) for quarterly data.}

\item{iter}{logical, \code{TRUE} (default) to conduct the boosted HP filter.
FALSE does not iterated, which is exactly the original HP filter.}

\item{stopping}{stopping criterion. \code{"BIC"} (default), or \code{"adf"}, or \code{"nonstop"} means keeping
iteration until the maximum number of iteration, specified by \code{Max_Iter} is reached.}

\item{sig_p}{a threshold of the p-value for the ADF test, with default value 0.050.
Only effective when \code{stopping = "adf"}.}

\item{Max_Iter}{maximal number of iterations. The default is 100.}
}
\value{
The function returns a list containing the following items:
\item{cycle}{The cyclical component in the final iteration.}
\item{trend}{The trend component in the final iteration.}
\item{trend_hist}{The estimated trend in each iteration.}
\item{iter_num}{The total number of iterations when it stops.}
\item{IC_hist}{The path of the BIC up to the final iterations.}
\item{adf_p_hist}{The path of the ADF test p-value up to the final iteration.}
}
\description{
All in one function of conducting the boosted HP-filter.
}
\details{
This is the main function of implementing the boosted HP filter (Phillisp and
Shi, 2021). The arguments accommendate the orginal HP filter (\code{iter =
FALSE}), the boosted HP filter with the BIC stopping criterion (\code{stopping =
"BIC"}),
or ADF test stopping criterion
(\code{stopping = "adf"}),  or keep going until the maximum number of iterations is reached
(\code{stopping = "nonstop"}).

Either the original HP filter or the bHP filter requires \code{lambda} to
control the strength of the weak learner for in-sample fitting. The default
is \code{lambda = 1600}, which is recommended by Hodrick and Prescott (1997)
for quarterly data. \code{lambda} should be adjusted for different
frequencies. For example, \code{lambda = 129600} for monthly data and
\code{lambda = 6.25} for annual data.

See the vignette with a brief introduction of the idea of bHP.
}
\examples{

data(IRE) # load the data 'IRE'
lam <- 100 # tuning parameter for the annual data

# raw HP filter
bx_HP <- BoostedHP(IRE, lambda = lam, iter= FALSE)

# by BIC
bx_BIC <- BoostedHP(IRE, lambda = lam, iter= TRUE, stopping = "BIC")

# by ADF
bx_ADF <- BoostedHP(IRE, lambda = lam, iter= TRUE, stopping = "adf")

# If stopping = "nonstop",
# Iterated HP filter until Max_Iter and keep the path of BIC.

bx_nonstop <- BoostedHP(IRE, lambda = lam, iter= TRUE, stopping = "nonstop")
}
\references{
Peter Phillips and Zhentao Shi, 2021: "Boosting: Why You Can Use the HP Filter," International Economic Review, 62(2), 521-570
}
