% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoostedHP.R
\name{BoostedHP}
\alias{BoostedHP}
\title{one in all function of iterated HP-filter}
\usage{
BoostedHP(x, lambda = 1600, iter = TRUE, test_type = "none",
  sig_p = 0.05, Max_Iter = 100)
}
\arguments{
\item{Max_Iter}{}
}
\value{
cycle,trend_hist,iter_num,IC_hist,adf_p_hist.
}
\description{
one in all function of iterated HP-filter
}
\examples{
bx_HP = BoostedHP(x, lambda = 1600, iter= FALSE)$trend
}
