% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bHP.R
\docType{data}
\name{IRE}
\alias{IRE}
\title{Ireland Annual GDP}
\format{\itemize{
 \item\strong{Release:} {Gross Domestic Product}
 \item\strong{Frequency:} {Annual}
 \item\strong{Date Range:} {1981--2016}
}}
\source{
OECD Stat \url{https://stats.oecd.org/}
}
\usage{
data(IRE)
}
\description{
Ireland Annual GDP
}
\section{Descripton}{

This dataset is described in Section 4.1 of Philips and Shi (2019).
Also See Okun, Ball, Leigh, and Loungani (2017).
}

\references{
Phillips, Peter CB, and Zhentao Shi.
"Boosting: Why you can use the hp filter."
arXiv: 1905.00175, Cowles Foundation Discussion Paper No.2192, (2019).

Ball, Laurence, Daniel Leigh, and Prakash Loungani.
  "Okun's law: Fit at 50?."
  Journal of Money, Credit and Banking 49, no. 7 (2017): 1413-1441.
}
\keyword{datasets}
