% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bHP.R
\docType{data}
\name{IRE}
\alias{IRE}
\title{Ireland Annual GDP}
\format{
\itemize{
 \item\strong{Release:} {Gross Domestic Product}
 \item\strong{Frequency:} {Annual}
 \item\strong{Date Range:} {1981--2016}
}
}
\source{
OECD Stat \url{https://stats.oecd.org/}
}
\usage{
data(IRE)
}
\description{
Ireland Annual GDP
}
\section{Description}{

This dataset is described in Section 4.1 of Philips and Shi (2019).
Also See Okun, Ball, Leigh, and Loungani (2017).
}

\references{
Peter Phillips and Zhentao Shi, 2021: "Boosting: Why You Can Use the HP Filter," International Economic Review, 62(2), 521-570

Ball, Laurence, Daniel Leigh, and Prakash Loungani.
  "Okun's law: Fit at 50?."
  Journal of Money, Credit and Banking 49, no. 7 (2017): 1413-1441.
}
\keyword{datasets}
